# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/user:
    get:
      tags: [Auth]
      operationId: getCurrentAuthUser
      summary: Get the currently authenticated user
      parameters:
        - $ref: "#/components/parameters/AuthorizationSessionHeader"
        - $ref: "#/components/parameters/SessionCookie"
      responses:
        "200":
          description: Auth status resolved successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUserEnvelope"
  /auth/register:
    post:
      tags: [Auth]
      operationId: registerAccount
      summary: Create an account (or claim one that has no password yet)
      description: >
        Creates a new account and signs it in. If an account with this email
        already exists but has no password (it predates the switch to
        email/password sign-in), the password is attached to that account so its
        scans, leaves and approval status are preserved. The first account ever
        created becomes the owner and is approved automatically; all others
        start unapproved.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Account created and signed in.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUserEnvelope"
        "400":
          description: Invalid email or password.
        "403":
          description: Registration code required or wrong.
        "409":
          description: An account with this email already exists.
  /auth/login:
    post:
      tags: [Auth]
      operationId: loginWithPassword
      summary: Sign in with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Signed in; the session cookie is set.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUserEnvelope"
        "401":
          description: Wrong email or password.
        "429":
          description: Too many failed attempts; try again later.
...
components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Signed session token — `Bearer <token>`.
      schema:
        type: string
    SessionCookie:
      name: auth
      in: cookie
      required: false
      description: Browser session cookie holding the signed token.
      schema:
        type: string
  schemas:
...
    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl, approved, isOwner, username, leavesCount]
      properties:
        id:
          type: string
        email:
          type: [string, 'null']
        firstName:
          type: [string, 'null']
        lastName:
          type: [string, 'null']
        profileImageUrl:
          type: [string, 'null']
        approved:
          type: boolean
          description: Whether the owner has granted this account access to the app.
        isOwner:
          type: boolean
          description: Whether this account is the app owner (the first account that ever signed in).
        username:
          type: [string, 'null']
        leavesCount:
          type: integer
    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: "#/components/schemas/AuthUser"
            - type: 'null'
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          minLength: 3
          description: Account email address (compared case-insensitively).
        password:
          type: string
          minLength: 8
          maxLength: 200
    RegisterInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          minLength: 3
        password:
          type: string
          minLength: 8
          maxLength: 200
        registrationCode:
          type: string
          description: >
            Shared invitation code. Only required when the server is configured
            with REGISTRATION_CODE.